refactor segmentation_info to roi_info
