+++ package/PartSeg/_roi_analysis/main_window.py
@@ -32,8 +32,8 @@ from PartSegCore.analysis.io_utils import create_history_element_from_project
 from PartSegCore.analysis.save_functions import save_dict
 from PartSegCore.io_utils import HistoryElement, WrongFileTypeException
 from PartSegCore.mask_create import calculate_mask_from_project
+from PartSegCore.roi_info import ROIInfo
 from PartSegCore.segmentation.algorithm_base import SegmentationResult
-from PartSegCore.segmentation_info import ROIInfo
 from PartSegImage import TiffImageReader
 
 from ..common_gui.algorithms_description import AlgorithmChoose, InteractiveAlgorithmSettingsWidget
